An idea to avoid reflection
This is one idea.  It uses package level access to make the constructors
and internal members only available to the builder in the enum.  The
only thing I dislike about this is the cast in the enum builder method.
This could be avoided by adding a "initialize" method to the interface,
but I didn't like that any better.